# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/mail_slots:
    get:
      tags:
        - MailSlots
      summary: List MailSlots
      description: Get a paginated list of MailSlots with optional filtering and sorting
      operationId: listMailSlots
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailSlotList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - MailSlots
      summary: Create MailSlot
      description: Create a new MailSlot
      operationId: createMailSlot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMailSlotRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailSlot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/registration_answers:
    get:
      tags:
        - RegistrationAnswers
      summary: List RegistrationAnswers
      description: Get a paginated list of RegistrationAnswers with optional filtering and sorting
      operationId: listRegistrationAnswers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationAnswerList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - RegistrationAnswers
      summary: Create RegistrationAnswer
      description: Create a new RegistrationAnswer
      operationId: createRegistrationAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRegistrationAnswerRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationAnswer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    MailSlot:
      type: object
      required:
        - id
        - scheduler_id
        - slot_id
        - mail_done
        - metadata
      properties:
        id:
          type: string
          format: uuid
        scheduler_id:
          type: string
          format: uuid
        slot_id:
          type: string
          format: uuid
        scheduled_date:
          nullable: true
          type: string
          format: date-time
        mail_done:
          type: boolean
        last_registration_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateMailSlotRequest:
      type: object
      required:
        - scheduler_id
        - slot_id
        - mail_done
        - metadata
      properties:
        scheduler_id:
          type: string
          format: uuid
        slot_id:
          type: string
          format: uuid
        scheduled_date:
          nullable: true
          type: string
          format: date-time
        mail_done:
          type: boolean
        last_registration_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    MailSlotList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MailSlot'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    RegistrationAnswer:
      type: object
      required:
        - id
        - registration_id
        - question_id
        - metadata
      properties:
        id:
          type: string
          format: uuid
        registration_id:
          type: string
          format: uuid
        question_id:
          type: string
          format: uuid
        value_text:
          nullable: true
          type: string
          maxLength: 2000
        value_answer_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateRegistrationAnswerRequest:
      type: object
      required:
        - registration_id
        - question_id
        - metadata
      properties:
        registration_id:
          type: string
          format: uuid
        question_id:
          type: string
          format: uuid
        value_text:
          nullable: true
          type: string
          maxLength: 2000
        value_answer_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    RegistrationAnswerList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RegistrationAnswer'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []