# src/presentation/http/mail_slot_handler.rs
//! MailSlot REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{MailSlotService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateMailSlotDto, UpdateMailSlotDto, PatchMailSlotDto, MailSlotResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum MailSlotError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for MailSlotError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for MailSlotError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "MAILSLOT_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "MAILSLOT_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "MAILSLOT_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "MAILSLOT_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for MailSlot.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /mail_slots | List with pagination |
/// | POST | /mail_slots | Create new |
/// | GET | /mail_slots/:id | Get by ID |
/// | PUT | /mail_slots/:id | Full update |
/// | PATCH | /mail_slots/:id | Partial update |
/// | DELETE | /mail_slots/:id | Soft delete |
/// | POST | /mail_slots/bulk | Bulk create |
/// | POST | /mail_slots/upsert | Upsert |
/// | GET | /mail_slots/trash | List deleted |
/// | POST | /mail_slots/:id/restore | Restore |
/// | DELETE | /mail_slots/empty | Empty trash |
/// | GET | /mail_slots/:id/deleted | Get deleted by ID |
/// | DELETE | /mail_slots/trash/:id | Permanent delete from trash |
/// | GET | /mail_slots/count | Count active entities |
/// | GET | /mail_slots/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(MailSlotService::with_repository(repository));
/// let router = create_mail_slot_routes(service);
/// ```
pub fn create_mail_slot_routes(service: Arc<MailSlotService>) -> Router {
    BackboneCrudHandler::<MailSlotService, MailSlot, CreateMailSlotDto, UpdateMailSlotDto, MailSlotResponseDto>::routes(
        service,
        "/mail_slots",
    )
}

/// Create Axum router with only the read (GET) endpoints for MailSlot.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_mail_slot_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_mail_slot_read_routes(service: Arc<MailSlotService>) -> Router {
    BackboneCrudHandler::<MailSlotService, MailSlot, CreateMailSlotDto, UpdateMailSlotDto, MailSlotResponseDto>::read_routes(
        service,
        "/mail_slots",
    )
}

/// Create Axum router with only the write (mutation) endpoints for MailSlot.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_mail_slot_write_routes(service: Arc<MailSlotService>) -> Router {
    BackboneCrudHandler::<MailSlotService, MailSlot, CreateMailSlotDto, UpdateMailSlotDto, MailSlotResponseDto>::write_routes(
        service,
        "/mail_slots",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_mail_slot_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<MailSlotService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_mail_slot_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}